+++ data-migrations/2024-09-25-1120-create-questions.sql
@@ -17,6 +17,7 @@ CREATE TABLE `questions` (
   `modified` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (templateId) REFERENCES templates(id) ON DELETE CASCADE,
   FOREIGN KEY (sectionId) REFERENCES sections(id) ON DELETE CASCADE,
+  FOREIGN KEY (questionTypeId) REFERENCES questionTypes(id),
   FOREIGN KEY (createdById) REFERENCES users(id),
   FOREIGN KEY (modifiedById) REFERENCES users(id),
   INDEX questions_section_idx (`sectionId`, `displayOrder`),