+++ src/resolvers/question.ts
@@ -2,9 +2,9 @@ import { Resolvers } from "../types";
 import { MyContext } from "../context";
 import { QuestionOption } from "../models/QuestionOption";
 import { Question } from "../models/Question";
+import { Template } from "../models/Template";
 import { getQuestionOptionsToRemove } from "../services/questionService";
 import { AuthenticationError, ForbiddenError, InternalServerError, NotFoundError } from "../utils/graphQLErrors";
-import { markTemplateAsDirty } from "../utils/templateUtils";
 import { QuestionCondition } from "../models/QuestionCondition";
 import { formatLogMessage } from "../logger";
 import { isAdmin, isAuthorized } from "../services/authService";
@@ -109,7 +109,7 @@ export const resolvers: Resolvers = {
             }
 
             // Update the associated template to set isDirty=1
-            await markTemplateAsDirty('Question resolver - addQuestion', context, templateId);
+            await Template.markTemplateAsDirty('Question resolver - addQuestion', context, templateId);
 
             // Return newly created question
             return await Question.findById(reference, context, questionId);
@@ -264,7 +264,7 @@ export const resolvers: Resolvers = {
           }
 
           // Update the associated template to set isDirty=1
-          await markTemplateAsDirty('Question resolver - updateQuestion', context, questionData.templateId);
+          await Template.markTemplateAsDirty('Question resolver - updateQuestion', context, questionData.templateId);
 
           // Refetch the question or the updated question with errors
           return updatedQuestion.hasErrors() ? updatedQuestion : await Question.findById(reference, context, questionId);
@@ -296,7 +296,7 @@ export const resolvers: Resolvers = {
           const question = new Question({ ...questionData, id: questionId });
 
           // Update the associated template to set isDirty=1
-          await markTemplateAsDirty('Question resolver - removeQuestion', context, questionData.templateId);
+          await Template.markTemplateAsDirty('Question resolver - removeQuestion', context, questionData.templateId);
 
           // The delete will also delete all associated questionOptions
           return await question.delete(context);